pyproject.toml
[tool.poetry]
name = "graphrag-demo"
version = "0.1.0"
description = "demo project for graphRAG"
authors = ["lishuai <lishuai.jing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.1"
openai = "^1.50.2"
neo4j = "^5.25.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.1"
tiktoken = "^0.7.0"
graphdatascience = "^1.11"
pyarrow = "15.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
